Don't close 'Edit Permissions' dialog without warning, if something is not saved #5120
Added confirmation dialog to the `EditPermissionsDialog`.
Moved `isDirty` function to `AppHelper`.
 Fixed focus blinking in dialogs with a confirmation dialog, when it was shown.
